Menu error if the data is more than 300 record in database.

+++ po-includes/core/sql/popojicms.sql
@@ -231,8 +231,8 @@ INSERT INTO `language` (`id_language`, `title`, `code`, `active`) VALUES
 
 DROP TABLE IF EXISTS `menu`;
 CREATE TABLE IF NOT EXISTS `menu` (
-  `id` tinyint(3) unsigned NOT NULL AUTO_INCREMENT,
-  `parent_id` tinyint(3) unsigned NOT NULL DEFAULT '0',
+  `id` smallint(5) unsigned NOT NULL AUTO_INCREMENT,
+  `parent_id` smallint(5) unsigned NOT NULL DEFAULT '0',
   `title` varchar(255) NOT NULL DEFAULT '',
   `url` varchar(255) NOT NULL DEFAULT '',
   `class` varchar(255) NOT NULL DEFAULT '',
